librbd: use check_exists
Optimistically assume the OSDs support CHECK_EXISTS op.  If they do, we
get a single op that attemps the write or falls back to copy-up.  If they
do not, we *then* send a STAT, and *then* (if that succeeds) send a
write.  So best case goes from 2 to 1 op, worst case goes from 2 to 3 ops.

I'm not entirely happy witht this.

Signed-off-by: Sage Weil <sage@inktank.com>