osdc/ObjectCacher: split lru evict and obj lclose
Allow objects to be removed from the lru cache if they are
trimmed and we don't hold any references to them, and let
those objects get closed later (by the commit).

This fixes a race between a trim of an lru'ed object
by readx and an osd commit reply.  If the object is
evicted first, we hit the assertion in close: assert(can_close()).

Signed-off-by: Sam Lang <sam.lang@inktank.com>