test: run w/ and w/o read from replica enabled
This patch does two things. First, it uses value-parameterized version
of the test fixture which allows us to pass in a boolean flag that
enables/disable read from replica feature. We create two instances of
the fixture (true and false), so we run all the tests twice, once for
each state.

Existing tests use getpid() to create unique file names, but often don't
clean up after themselves. Since each tests is run in the same process
more than once, the names are no longer unique to the test instance. To
handle this case we change the fixture to construct a unique directory
that the mount is rooted in to create a clean room directory for each
test.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>